# src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Logbrook API",
    "version": "0.1.0"
  },
  "paths": {
    "/logs": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "description": "Inclusive Unix milliseconds",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "description": "Exclusive Unix milliseconds",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "description": "Comma-separated authenticated source scope; omitted means token scope, empty configured scope is unrestricted",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service",
            "in": "query",
            "required": false,
            "description": "Exact service",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "logger",
            "in": "query",
            "required": false,
            "description": "Exact logger",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "host",
            "in": "query",
            "required": false,
            "description": "Exact host",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "message",
            "in": "query",
            "required": false,
            "description": "Literal message substring",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_level",
            "in": "query",
            "required": false,
            "description": "Minimum numeric severity",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size, 1 to configured maximum",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque stable pagination cursor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Source outside read scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds configured budget",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Capacity exhausted; Retry-After: 1",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Query deadline expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/logs/ingest": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Source outside read scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds configured budget",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Capacity exhausted; Retry-After: 1",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Query deadline expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "507": {
            "description": "Index physical size budget exhausted; oldest retained history may be reclaimed before subsequent admission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Atomic live ingestion: event timestamps must lie between now minus configured retention and now plus configured maximum future skew (default five minutes). Every unconsumed producer key is preserved in attributes, including canonical-name collisions.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "minItems": 1,
                "items": {
                  "$ref": "#/components/schemas/PinoEvent"
                }
              }
            },
            "application/x-ndjson": {
              "schema": {
                "type": "string"
              },
              "example": "{\"time\":1735689600000,\"level\":30,\"msg\":\"hello\"}\n"
            }
          }
        }
      }
    },
...
    "/indexes/{index}/logs": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "index",
            "in": "path",
            "required": true,
            "description": "Named index; lowercase letters, digits, underscores and hyphens. Index token scope applies in addition to source scope.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 63,
              "pattern": "^[a-z0-9][a-z0-9_-]*$"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "description": "Inclusive Unix milliseconds",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "description": "Exclusive Unix milliseconds",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "description": "Comma-separated authenticated source scope; omitted means token scope, empty configured scope is unrestricted",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service",
            "in": "query",
            "required": false,
            "description": "Exact service",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "logger",
            "in": "query",
            "required": false,
            "description": "Exact logger",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "host",
            "in": "query",
            "required": false,
            "description": "Exact host",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "message",
            "in": "query",
            "required": false,
            "description": "Literal message substring",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_level",
            "in": "query",
            "required": false,
            "description": "Minimum numeric severity",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size, 1 to configured maximum",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque stable pagination cursor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Source outside read scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds configured budget",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Capacity exhausted; Retry-After: 1",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Query deadline expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Index not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Named route cursors and SSE Last-Event-ID are opaque and bound to this index. Authenticated ingestion may create a missing index within max_indexes; reads never create indices."
      }
    },
...
    "/indexes/{index}/logs/ingest": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "index",
            "in": "path",
            "required": true,
            "description": "Named index; lowercase letters, digits, underscores and hyphens. Index token scope applies in addition to source scope.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 63,
              "pattern": "^[a-z0-9][a-z0-9_-]*$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Source outside read scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds configured budget",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Capacity exhausted; Retry-After: 1",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Query deadline expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Index not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "507": {
            "description": "Index physical size budget exhausted; oldest retained history may be reclaimed before subsequent admission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Atomic live ingestion: event timestamps must lie between now minus configured retention and now plus configured maximum future skew (default five minutes). Every unconsumed producer key is preserved in attributes, including canonical-name collisions. Named route cursors and SSE Last-Event-ID are opaque and bound to this index. Authenticated ingestion may create a missing index within max_indexes; reads never create indices.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "minItems": 1,
                "items": {
                  "$ref": "#/components/schemas/PinoEvent"
                }
              }
            },
            "application/x-ndjson": {
              "schema": {
                "type": "string"
              },
              "example": "{\"time\":1735689600000,\"level\":30,\"msg\":\"hello\"}\n"
            }
          }
        }
      }
    },
...
    "/indexes": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "List existing indexes authorized by a read token index scope, or all indexes with an admin token. Source-scoped read tokens receive names only. Other authorized callers receive measured file sizes and policy.",
        "responses": {
          "200": {
            "description": "Authorized indices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "indexes"
                  ],
                  "properties": {
                    "indexes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/IndexInfo"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Read or admin token required"
          },
          "429": {
            "description": "Diagnostics capacity exhausted"
          },
          "503": {
            "description": "Filesystem inspection unavailable"
          }
        }
      }
    },
    "/indexes/{index}": {
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "index",
            "in": "path",
            "required": true,
            "description": "Named index; lowercase letters, digits, underscores and hyphens. Index token scope applies in addition to source scope.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 63,
              "pattern": "^[a-z0-9][a-z0-9_-]*$"
            }
          }
        ],
        "description": "Create an index using the configured policy. Requires an admin token, or ingestion token with index scope. Idempotently returns an existing index; global max_indexes bounds creation.",
        "responses": {
          "200": {
            "description": "Index exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "name"
                  ],
                  "properties": {
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Ingestion or admin token required"
          },
          "403": {
            "description": "Index outside token scope"
          },
          "429": {
            "description": "Capacity exhausted"
          }
        }
      },
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "index",
            "in": "path",
            "required": true,
            "description": "Named index; lowercase letters, digits, underscores and hyphens. Index token scope applies in addition to source scope.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 63,
              "pattern": "^[a-z0-9][a-z0-9_-]*$"
            }
          }
        ],
        "description": "Inspect an existing index. A read token must permit the index; an admin token permits every index. Source-scoped readers receive only the name. Other authorized callers receive fresh file sizes and effective retention policy.",
        "responses": {
          "200": {
            "description": "Index information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexInfo"
                }
              }
            }
          },
          "401": {
            "description": "Read or admin token required"
          },
          "403": {
            "description": "Index outside token scope"
          },
          "404": {
            "description": "Unknown index"
          },
          "429": {
            "description": "Diagnostics capacity exhausted"
          },
          "503": {
            "description": "Filesystem inspection unavailable"
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "index",
            "in": "path",
            "required": true,
            "description": "Named index; lowercase letters, digits, underscores and hyphens. Index token scope applies in addition to source scope.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 63,
              "pattern": "^[a-z0-9][a-z0-9_-]*$"
            }
          }
        ],
        "description": "Permanently delete a dynamic index and its events. Requires a separate admin token. The default index and indexes declared in server configuration are protected. An admitted deletion drains native workers and finishes independently of client disconnection. Producers can recreate a deleted index on subsequent ingestion; stop producers first when removing an index permanently.",
        "responses": {
          "200": {
            "description": "Index deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "deleted"
                  ],
                  "properties": {
                    "deleted": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid name or protected default/configured index"
          },
          "401": {
            "description": "Valid credential required"
          },
          "403": {
            "description": "Admin token required"
          },
          "404": {
            "description": "Unknown index"
          },
          "429": {
            "description": "Request capacity exhausted"
          },
          "503": {
            "description": "Shutdown or filesystem operation failed; inspect state before retrying"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "description": "Configured tokens contain 16 to 512 visible ASCII characters without whitespace.",
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "PinoEvent": {
        "type": "object",
        "required": [
          "time",
          "level",
          "msg"
        ],
        "properties": {
          "time": {
            "type": "integer",
            "minimum": 0,
            "description": "Unix milliseconds"
          },
          "level": {
            "type": "integer",
            "minimum": 0
          },
          "msg": {
            "type": "string"
          },
          "service": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "hostname": {
            "type": [
              "string",
              "null"
            ]
          },
          "pid": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          }
        },
        "additionalProperties": true
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "SearchResult": {
        "type": "object",
        "required": [
          "events",
          "next_cursor"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Accepted": {
        "type": "object",
        "required": [
          "accepted"
        ],
        "properties": {
          "accepted": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    }
  }
}